chore(app-server): split control contract crate
