+++ Cargo.toml
@@ -17,6 +17,7 @@ members = [
     "crates/rara-observability",
     "crates/rara-tools",
     "crates/rara-plugins",
+    "crates/rara-app-server",
 ]
 resolver = "2"
 
@@ -53,6 +54,7 @@ rara-memory = { path = "crates/rara-memory" }
 rara-observability = { path = "crates/rara-observability" }
 rara-tools = { path = "crates/rara-tools" }
 rara-plugins = { path = "crates/rara-plugins" }
+rara-app-server = { path = "crates/rara-app-server" }
 rara-state = { path = "crates/rara-state" }
 tokio = { version = "1.0", features = ["full"] }
 tokio-util = { version = "0.7", features = ["compat"] }
