+++ crates/rara-app-server/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "rara-app-server"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+serde = { workspace = true }
+serde_json = { workspace = true }
