+++ crates/rara-app-server/src/lib.rs
@@ -1,0 +1,3 @@
+//! App-server protocol contracts shared by transport adapters and the runtime.
+
+pub mod runtime_control;
