+++ src/runtime_control.rs
@@ -1,3 +1,11 @@
+pub use rara_app_server::runtime_control::{
+    ApprovalControlRequest, HookControlRequest, HookLifecycle, InputControlRequest,
+    McpControlRequest, MemoryControlRequest, MemoryRecordControlPatch, MemoryScope,
+    OutputSubscriptionRequest, PromptSourceControlRequest, PromptSourceLifetime,
+    PromptSourceRegistration, RuntimeControlEnvelope, RuntimeControlRequest, RuntimeControllerKind,
+    RuntimeProvenance, RuntimeSourceAuthorship, RuntimeSourceTrust, SessionControlRequest,
+    ShellApprovalDecision, SkillSourceControlRequest, SourceLayer, SourceScope,
+};
 use rara_tools::tool::ToolOutputStream;
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
@@ -7,142 +15,6 @@ use crate::context::{ContextObservabilityView, RetrievalOrchestrationView};
 use crate::mcp_status::{McpConnectionState, McpStatusSnapshot};
 use crate::session_promotion::SessionShardPromotionOutcome;
 use crate::todo::TodoState;
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum RuntimeControllerKind {
-    LocalTui,
-    LocalCli,
-    Acp,
-    Wire,
-    AppServer,
-    Runtime,
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct RuntimeProvenance {
-    pub controller: RuntimeControllerKind,
-    pub adapter: Option<String>,
-    pub session_id: Option<String>,
-    pub source_id: Option<String>,
-    pub trust: RuntimeSourceTrust,
-    pub authorship: RuntimeSourceAuthorship,
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum RuntimeSourceTrust {
-    Trusted,
-    Untrusted,
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum RuntimeSourceAuthorship {
-    UserProvided,
-    Generated,
-    Runtime,
-}
-
-#[allow(dead_code)]
-impl RuntimeProvenance {
-    pub fn local_tui(session_id: impl Into<String>) -> Self {
-        Self {
-            controller: RuntimeControllerKind::LocalTui,
-            adapter: None,
-            session_id: Some(session_id.into()),
-            source_id: None,
-            trust: RuntimeSourceTrust::Trusted,
-            authorship: RuntimeSourceAuthorship::UserProvided,
-        }
-    }
-
-    pub fn runtime(session_id: Option<String>) -> Self {
-        Self {
-            controller: RuntimeControllerKind::Runtime,
-            adapter: None,
-            session_id,
-            source_id: None,
-            trust: RuntimeSourceTrust::Trusted,
-            authorship: RuntimeSourceAuthorship::Runtime,
-        }
-    }
-
-    pub fn protocol(
-        controller: RuntimeControllerKind,
-        adapter: impl Into<String>,
-        session_id: Option<String>,
-        source_id: Option<String>,
-    ) -> Self {
-        Self {
-            controller,
-            adapter: Some(adapter.into()),
-            session_id,
-            source_id,
-            trust: RuntimeSourceTrust::Untrusted,
-            authorship: RuntimeSourceAuthorship::UserProvided,
-        }
-    }
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-#[serde(tag = "type", content = "payload", rename_all = "snake_case")]
-pub enum RuntimeControlRequest {
-    Session(SessionControlRequest),
-    Input(InputControlRequest),
-    Output(OutputSubscriptionRequest),
-    PromptSource(PromptSourceControlRequest),
-    SkillSource(SkillSourceControlRequest),
-    Mcp(McpControlRequest),
-    Memory(MemoryControlRequest),
-    Hook(HookControlRequest),
-    Approval(ApprovalControlRequest),
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-pub struct RuntimeControlEnvelope {
-    pub request_id: String,
-    pub provenance: RuntimeProvenance,
-    pub request: RuntimeControlRequest,
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "type", content = "payload", rename_all = "snake_case")]
-pub enum SessionControlRequest {
-    CreateSession,
-    ResumeSession { session_id: String },
-    CancelCurrentTurn,
-    InterruptCurrentTurn,
-    QueryRuntimeState,
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "type", content = "payload", rename_all = "snake_case")]
-pub enum InputControlRequest {
-    SubmitUserPrompt { prompt: String },
-    AnswerPendingInput { answer: String },
-    AnswerPlanApproval { approved: bool },
-    AnswerShellApproval { decision: ShellApprovalDecision },
-    SubmitFollowUp { prompt: String },
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum ShellApprovalDecision {
-    Once,
-    Prefix,
-    Always,
-    Suggestion,
-}
 
 impl From<BashApprovalDecision> for ShellApprovalDecision {
     fn from(decision: BashApprovalDecision) -> Self {
@@ -164,177 +36,6 @@ impl From<ShellApprovalDecision> for BashApprovalDecision {
             ShellApprovalDecision::Suggestion => Self::Suggestion,
         }
     }
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "type", content = "payload", rename_all = "snake_case")]
-pub enum OutputSubscriptionRequest {
-    Subscribe { subscriber_id: String },
-    Unsubscribe { subscriber_id: String },
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "type", content = "payload", rename_all = "snake_case")]
-pub enum PromptSourceLifetime {
-    Turns(u32),
-    Session,
-    Persistent,
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-pub struct PromptSourceRegistration {
-    pub source_id: String,
-    pub scope: SourceScope,
-    pub layer: SourceLayer,
-    pub budget_hint_tokens: Option<u32>,
-    pub lifetime: PromptSourceLifetime,
-    pub content: String,
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum SourceScope {
-    Home,
-    Repo,
-    CurrentWorkingDirectory,
-    Session,
-    Protocol,
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum SourceLayer {
-    System,
-    Developer,
-    User,
-    Memory,
-    Skill,
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "type", content = "payload", rename_all = "snake_case")]
-pub enum PromptSourceControlRequest {
-    Register(PromptSourceRegistration),
-    Unregister { source_id: String },
-    QuerySources,
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "type", content = "payload", rename_all = "snake_case")]
-pub enum SkillSourceControlRequest {
-    RegisterRoot {
-        source_id: String,
-        root: String,
-        precedence_hint: Option<i32>,
-    },
-    RegisterSkill {
-        source_id: String,
-        name: String,
-        content: String,
-        precedence_hint: Option<i32>,
-    },
-    DisableSkill {
-        name: String,
-        source_id: Option<String>,
-    },
-    QuerySkills,
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "type", content = "payload", rename_all = "snake_case")]
-pub enum McpControlRequest {
-    QueryStatus,
-    Refresh { server_name: Option<String> },
-    Reconnect { server_name: String },
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
-#[serde(tag = "type", content = "payload", rename_all = "snake_case")]
-pub enum MemoryControlRequest {
-    AddRecord {
-        memory_id: String,
-        scope: MemoryScope,
-        content: String,
-        metadata: Value,
-    },
-    UpdateRecord {
-        memory_id: String,
-        patch: MemoryRecordControlPatch,
-    },
-    DeleteRecord {
-        memory_id: String,
-    },
-    ListLabels {
-        scope: Option<MemoryScope>,
-    },
-    QueryRecords {
-        query: String,
-        scope: Option<MemoryScope>,
-        limit: usize,
-    },
-    QueryMetadata,
-    SelectionSnapshot,
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, Default, PartialEq, Serialize, Deserialize)]
-pub struct MemoryRecordControlPatch {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub title: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub content: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub labels: Option<Vec<String>>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub importance: Option<f64>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub pinned: Option<bool>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub scope: Option<MemoryScope>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub session_id: Option<Option<String>>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub thread_id: Option<Option<String>>,
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(rename_all = "snake_case")]
-pub enum MemoryScope {
-    Thread,
-    Workspace,
-}
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "type", content = "payload", rename_all = "snake_case")]
-pub enum HookControlRequest {
-    Declare {
-        hook_id: String,
-        lifecycle: HookLifecycle,
-        description: String,
-    },
-    QueryHooks,
-}
-
-#[allow(dead_code)]
-pub use rara_instructions::HookLifecycle;
-
-#[allow(dead_code)]
-#[derive(Clone, Debug, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(tag = "type", content = "payload", rename_all = "snake_case")]
-pub enum ApprovalControlRequest {
-    AnswerPendingApproval { approval_id: String, approved: bool },
-    QueryPendingApprovals,
 }
 
 #[allow(dead_code)]
